+++ examples/custom/bindings/entry_main.c
@@ -1,17 +1,28 @@
 #include <wisdom.h>
 #include <stdio.h>
 
+void ExDebugCallback(WisSeverity sev, const char* message, void* user_data)
+{
+    printf("Debug: %s\n", message);
+}
+
 int main()
 {
     WisFactory factory = NULL;
     WisDevice device = NULL;
+    WisDebugMessenger debug = NULL;
 
     WisFactoryExtQuery exts[] = {
         { FactoryExtIDDebugExtension, NULL },
     };
 
     WisCreateFactory(true, exts, 1, &factory);
+    WisDebugExtension debug_ext = exts[0].result;
 
+    // Get debug messenger
+    WisDebugExtensionCreateDebugMessenger(debug_ext, &ExDebugCallback, NULL, &debug);
+
+    // Create a device
     for (size_t i = 0;; i++) {
         WisAdapter adapter = NULL;
         WisResult result = WisFactoryGetAdapter(factory, i, AdapterPreferencePerformance, &adapter);